Add files via upload
 A versatile application that contains, as you will see, 12 different applications, some useful and practical for everyday use, some for fun.

Brief description of some of them, the others are self-explanatory, like all of them. But I think it's useful to have a first idea of what it's about.

Before I start the short descriptions, let me first tell you about the menu buttons, they change and appear by swiping left and right.

1η)  

Calendar

A complete calendar to schedule the meetings you want to have

2η)

Make a list:

With it, you can register and then log in to the database where you can write something   that you think might be useful.

3η)

Sharp memory

A very simple game of finding 2 sane numbers. After you complete it by finding all the numbers, at the end you have two options. One with the replay, to repeat the game with the numbers in the same boxes and with the other option, the numbers are placed in different random positions.

4η)

Find address

With this app, you can put any address you have and find it on the map with the route to it.

5η)

Where is my car

With it you can find your car where you have parked it, just press the button before you go away from your car:: Remember my current location

6η)

Personal assistant

 All you do is press the microphone button on the bottom and just voice ask for what you normally search for in navigators

7η)

Dice game

With it you just press the roll a good dice button and it is played with the dice that keep appearing in a completely random order and random combination.

8η)

Music

With it you hear improvised music that I play and that appear in YouTube videos

9η)

Share photos

With it, you can share photos or videos you take with your mobile phone with anyone you want,  with  email , or in social media.

10η)

Calendar

A different calendar you spend what you want to plan for the future, or even what you want to remember from past dates.

11η)

Compass

A compass for whether you find yourself in the middle of the ocean or the jungle

12η)

Shopping List

With it you write the shopping of the day or the week, with the possibility to account for the expenses you made for shopping you bought, selecting what you bought and leaving the selection blank for what you didn't buy for some reason.

Save the selected purchases, after the button: Finish shopping, you can share the purchases you made, for example, with your partner so that he or she knows what you left out of the market, and even how much you spent on those purchases.

These in short, you will find most of them by searching for the application

Made  by Evangelos Rizopoulos

Μία πολύπτυχη εφαρμογή που περιέχει όπως θα δείτε 12 διαφορετικές εφαρμογές, κάποιες χρήσιμες και πρακτικές για χρήση στην καθημερινότητα, κάποιες για διασκέδαση.

Περιληπτική περιγραφή κάποιων από αυτές, οι άλλες είναι αυτονόητε κατανοητές, όπως όλες. Αλλά θεωρώ χρήσιμο να έχετε μια πρώτη εικόνα, περι τίνος πρόκειται.

1η)

Calendar

Ένα πλήρες ημερολόγιο για να προγραμματίσετε τις συναντήσεις που θέλετε να έχετε 

2η)

Make a list:

Με αυτήν μπορείτε να κάνετε εγγραφή και είσοδο μετά στην βάση δεδομένων όπου μπορείτε να γράφετε κάτι περιληπτικό που θεωρείται ότι μπορεί να σας είναι χρήσιμο.

3η)

Sharp memory

Ένα πολύ απλό παιχνίδι εύρεσης 2 όμοιων αριθμών. Αφού το ολοκληρώσετε βρίσκοντας όλους τους αριθμούς, στο τέλος έχετε δύο επιλογές. Η μια με το replay  να επαναλάβετε το παιχνίδι με τους αριθμούς στα ίδια πλαίσια και με την άλλη επιλογή, η αριθμοί να τοποθετηθούν σε διαφορετικές τυχαίες θέσεις.

4η)

Find address

Με αυτήν την εφαρμογή, μπορείτε να βάζετε μια οποιαδήποτε διεύθυνση έχετε και να την βρείτε στον χάρτη με την διαδρομή προς αυτήν.

5η)

Where is my car

Με αυτήν μπορείτε να βρίσκεται το αμάξι σας όπου το έχετε σταθμεύσει, αρκει να πατήσετε το κουμπί:  Remember my current location

6η)

Personal assistant

 Το μόνο που κάνετε είναι να πατήσετε το κουμπί του μικροφώνου που έχει στο κάτω μέρος και να ζητήσετε μόνο φωνητικά ό,τι ψάχνετε κανονικά στους πλοηγούς

7η)

Dice game

Με αυτήν απλά και μόνο πατάτε το κουμπί ρίξε μια ζαριά καλή και παίζεται με τα ζάρια που  εμφανίζονται συνέχεια με μια εντελώς τυχαία σειρά και τυχαίο συνδυασμό.

8η)

Music

Με αυτήν ακούτε αυτοσχιαδιαστική μουσική που παίζω εγώ και εμφανίζονται σε βίντεο του YouTube

9η)

Share photos

Με αυτήν μοιράζεστε με όποιον θέλετε φωτογραφίες ή βίντεο που τραβάτε με το κινητό σας.

10η)

Calendar

Ένα ημερολόγιο διαφορετικό περνάτε ότι θέλετε να προγραμματίσετε για προσεχώς, ή ακόμα και τι θέλετε να θυμάστε από παρελθούσες ημερομηνίες.

11η)

Compass

Μια πυξίδα για το αν βρεθείτε στο μέσον του ωκεανού ή της ζούγκλας

12η)

Shopping List

Με αυτήν γράφεται τα ψώνια της ημέρας ή της εβδομάδας, με δυνατότητα να λογαριάζετε τα έξοδα που κάνατε για ψώνια που αγοράσατε, επιλέγοντας όσα αγοράσατε και αφήνοντας κενό την επιλογή σε αυτά που δεν αγοράσατε για κάποιον λόγο.

Αποθηκεύετε τα επιλεγμένα που αγοράσατε, μετά το κουμπί: Finish shopping μπορείτε να μοιραστείτε τις αγορές κάνατε π.χ., με τον ή την σύντροφό για να ξέρει τι αφήσατε εκτός αγοράς, ακόμα και πόσα ξοδέψατε για αυτά τα ψώνια.

Αυτά εν ολίγοις, τα περισσότερα θα τα βρείτε ψάχνοντας την εφαρμογή

Κατασκευασθείσα από τον Ευάγγελο Ριζόπουλο